==> notebooks/2_ranking_text_generation_pipeline/0_get_amsterdamnl_only_questions.ipynb <==
%cd /Users/natalipeeva/Documents/GitHub/Automatic-Answering-of-City-Council-Questions/

import os

# Get the notebook directory
notebook_dir = os.getcwd()

# Get the root directory by navigating upwards two levels
root_dir = os.path.dirname(os.path.abspath(os.path.join(notebook_dir, '../')))

import sys
src_dir = os.path.join(root_dir, 'src')
sys.path.append(src_dir)

from read_data.read_data import read_urls_questions, get_questions, get_url_content_tuples, get_relevant_docs
import pandas as pd

import pandas as pd

collected, questions =  read_urls_questions(os.path.join(root_dir, 'data/all_references/reference_urls_collected.csv'),
                                            os.path.join(root_dir, 'data/all_references/questions_updated_urls.csv'), clean_url_nan=True) # read collected urls and questions + remove unsuccessful collection
amsterdam_only = questions[questions['Cleaned_URLs'].apply(lambda urls: any(url.startswith('https://www.amsterdam.nl') for url in urls))]
amsterdam_collected = pd.read_csv(os.path.join(root_dir, 'data/collected/combined.csv'))
amsterdam_references = collected[collected['URL'].str.startswith('https://www.amsterdam.nl')]
amsterdam_missing = pd.DataFrame()  # Empty DataFrame to store missing rows
missing_rows = []

for index, row in amsterdam_references.iterrows():
    if row['URL'] not in amsterdam_collected['URL'].tolist():
        missing_rows.append(row.to_dict())

amsterdam_missing = pd.DataFrame(missing_rows)
all_amsterdam = pd.concat([amsterdam_collected, amsterdam_missing])

### Save Data

amsterdam_only.to_csv(os.path.join(root_dir, 'data/amsterdam/amsterdam_questions.csv'))
all_amsterdam.to_csv(os.path.join(root_dir, 'data/amsterdam/amsterdam_full.csv'))
