==> notebooks/2_ranking_text_generation_pipeline/1_ranking_comparison_tfidf_random_bm25.ipynb <==
### Setup CD

%cd  /Users/natalipeeva/Documents/GitHub/Automatic-Answering-of-City-Council-Questions/

import sys
import os
src_dir = os.path.join(os.getcwd(), 'src_clean')
sys.path.append(src_dir)

import pandas as pd
from hashids import Hashids
from tabulate import tabulate

from src_clean.preprocessing.passages import create_passages_dataframe, add_passages_ids
from src_clean.preprocessing.text_preprocessing import preprocess_text
from src_clean.ranking.sparse_ranking import perform_tfidf_search, perform_random_search, perform_bm25_search
from src_clean.ranking.evaluation import calculate_average_metrics_retrieval, calculate_metrics_answer_similarity, simulate_answer

### Read Data

questions_for_ranking = pd.read_csv('data/amsterdam/amsterdam_questions.csv')
collection = pd.read_csv('data/amsterdam/amsterdam_full.csv')

## Prepare Data for ranking

### Create passages

passages_df = create_passages_dataframe(collection) # 50 secs for execution

### Clean passages

#### Drop duplicates in general; drop duplictates in column 'Textual_Content'

passages_df = passages_df.drop_duplicates() # drop dup.

passages_df = passages_df.drop_duplicates(subset='Textual_Content', keep=False) # drop dup. passages

#### Pre-process Passages Text

passages_df['Preprocessed_Text'] = passages_df['Textual_Content'].apply(lambda x: preprocess_text(x,stem=True,
                                                                                      remove_stopwords=True,
                                                                                      lowercase_text=True,
                                                                                      remove_punct=True)) # 40 seconds

#### Add Passage ids

hashids = Hashids()
passages_df["id"] = [hashids.encode(i) for i in range(len(passages_df))]

passages_df.tail()

#### Check how many URLs are matching after pre-processing
- the removing duplicates in particular might have caused some URLs to not be in the collection anymore, however, if there is relevant content, it should still be in the passage collection

import ast

urls_unformatted = list(questions_for_ranking['URLs'])
reference_list = []
for string in urls_unformatted:
    sublist = ast.literal_eval(string)
    reference_list.append(sublist)

reference_urls = []
for sublist in reference_list:
    reference_urls.extend(sublist)

collected_urls = list(passages_df['URL'])
count = 0
elements_not_in_collected = []
for element in reference_urls:
    if element in collected_urls:
        count += 1
    else:
        elements_not_in_collected.append(element)


#count # it was 80 before droping duplicates, so some URLs are not in the collection anymore

### Clean Questions

#### Pre-process questions and answers

questions_for_ranking['Preprocessed_Question'] = questions_for_ranking['Question'].apply(lambda x: preprocess_text(x,stem=True,
                                                                                      remove_stopwords=True,
                                                                                      lowercase_text=True,
                                                                                      remove_punct=True)) 

questions_for_ranking['Preprocessed_Answer'] = questions_for_ranking['Answer'].apply(lambda x: preprocess_text(x,stem=True,
                                                                                      remove_stopwords=True,
                                                                                      lowercase_text=True,
                                                                                      remove_punct=True)) 

#questions_for_ranking.head()

#### add relevant passage ids to questions 
- Meaning: add the passages that are part of referenced URLs

questions_for_ranking = add_passages_ids(questions_for_ranking, passages_df) # 1:17 min

# questions_for_ranking.head()

#### remove the questions that have no reference anymore 
- Idea: so it doesn't bring down retrieval scores + avoid zero division

questions_for_ranking = questions_for_ranking[questions_for_ranking['passages_ids'].notna()] # remove no reference

#len(questions_for_ranking) # 52 samples left out of around 70

##### Add question ids 

hashids = Hashids()
questions_for_ranking["question_id"] = [hashids.encode(i) for i in range(len(questions_for_ranking))]

# questions_for_ranking.tail() # check

#### remove unused columns - questions

#questions_for_ranking.drop('Unnamed: 0.1', axis=1, inplace=True)
#questions_for_ranking.drop('Unnamed: 0', axis=1, inplace=True)
#questions_for_ranking.drop('Cleaned_URLs', axis=1, inplace=True)

#### prepare dataset

from datasets import Dataset
import pyarrow as pa

arrow_table = pa.Table.from_pandas(questions_for_ranking)
arrow_dict = arrow_table.to_pydict()
questions_for_ranking = Dataset.from_dict(arrow_dict)

# questions_for_ranking[0] # check

arrow_table = pa.Table.from_pandas(passages_df)
arrow_dict = arrow_table.to_pydict()
passages_df = Dataset.from_dict(arrow_dict)

# passages_df[1000] # check

type(passages_df)

## Ranking

### TF-IDF search

results_tfidf = perform_tfidf_search(questions_for_ranking, passages_df, k=100, get_true_passages=True) # 1min:13 without NN

### Random

results_random = perform_random_search(questions_for_ranking, passages_df, k=100, get_true_passages=True) # 18 secs

### BM25

results_bm25 = perform_bm25_search(questions_for_ranking, passages_df, k=100, get_true_passages=True) # 1min 24

### Check Average Answer length
- Idea: replicate length with retrieved passages to 'simulate' an answer

from nltk.tokenize import word_tokenize
import nltk
# Set the language to Dutch
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')
nltk.download('tagsets')

ans_len = []
for q in questions_for_ranking:
    ans_len.append(len(word_tokenize(q['Answer'], language='dutch')))
sum(ans_len)/len(ans_len) # 256.87 -> make limit for tokens in ranked content 256

## Evaluation

### Retrieval Scores - ranking method

#### Make Table

ks = [5, 10, 100]
results_lists = [results_tfidf, results_random, results_bm25]
retrieval_names = ['TF-IDF', 'Random', 'BM25']

table = []

for results, retrieval_name in zip(results_lists, retrieval_names):
    for k in ks:
        row = ['Metrics for k={} (Retrieval: {})'.format(k, retrieval_name)]
        average_metrics = calculate_average_metrics_retrieval(results, k)
        for metric, value in average_metrics.items():
            row.append('{:.4f}'.format(value))
        table.append(row)
headers = [''] + list(average_metrics.keys())

#### Table

print(tabulate(table, headers=headers, tablefmt="grid"))

#### ROUGE Scores - Preprocessed ranked text and answer

#### Make Table

answers = list(questions_for_ranking['Preprocessed_Answer']) # preprocessed

results_lists = [results_tfidf, results_random, results_bm25]
retrieval_names = ['TF-IDF', 'Random', 'BM25']

simulated_answers = {}
metrics = {}

for results, retrieval_name in zip(results_lists, retrieval_names):
    simulated_answers[retrieval_name] = []
    metrics[retrieval_name] = {}
    
    for result in results: # generate simulated answers = get top retrieved documents till tokens are < 256
        simulated_answer = simulate_answer(result['ranked_text_preprocessed'])
        simulated_answers[retrieval_name].append(simulated_answer)
    
    metric_values = calculate_metrics_answer_similarity(simulated_answers[retrieval_name], answers) # get ROUGE 
    
    metrics[retrieval_name] = metric_values

# Print the metrics table
table = []
headers = ['Retrieval Method'] + list(metric_values.keys())

for retrieval_name in retrieval_names:
    row = [retrieval_name] + [metrics[retrieval_name][metric] for metric in metric_values.keys()]
    table.append(row)


#### Table

print(tabulate(table, headers=headers, tablefmt="grid"))

#### ROUGE Scores - Unpreprocessed ranked text and answer

#### Make table

answers = list(questions_for_ranking['Answer']) # unpreprocessed

results_lists = [results_tfidf, results_random, results_bm25]
retrieval_names = ['TF-IDF', 'Random', 'BM25']

simulated_answers = {}
metrics = {}

for results, retrieval_name in zip(results_lists, retrieval_names):
    simulated_answers[retrieval_name] = []
    metrics[retrieval_name] = {}
    
    for result in results: # generate simulated answers = get top retrieved documents till tokens are < 256
        simulated_answer = simulate_answer(result['ranked_text']) #unpreprocessed
        simulated_answers[retrieval_name].append(simulated_answer)
    
    metric_values = calculate_metrics_answer_similarity(simulated_answers[retrieval_name], answers) # get ROUGE 
    
    metrics[retrieval_name] = metric_values

# Print the metrics table
table = []
headers = ['Retrieval Method'] + list(metric_values.keys())

for retrieval_name in retrieval_names:
    row = [retrieval_name] + [metrics[retrieval_name][metric] for metric in metric_values.keys()]
    table.append(row)

#### Table

print(tabulate(table, headers=headers, tablefmt="grid"))

## Save ranked results

import os
import pickle

def save_results(results, question_collection, passages_collection, folder_path):
    # Create the folder if it doesn't exist
    os.makedirs(folder_path, exist_ok=True)
    
    # Get the variable names of question_collection and passages_collection
    question_var_name = [var_name for var_name, var_val in globals().items() if var_val is question_collection][0]
    passages_var_name = [var_name for var_name, var_val in globals().items() if var_val is passages_collection][0]

    # Save each result variable with its corresponding filename
    for result in results:
        # Get the variable name of the result variable
        result_var_name = [var_name for var_name, var_val in globals().items() if var_val is result][0]
        
        # Generate the filename based on variable names
        filename = '{}_{}_{}.pickle'.format(result_var_name, question_var_name, passages_var_name)
        
        file_path = os.path.join(folder_path, filename)
        with open(file_path, 'wb') as f:
            pickle.dump(result, f)


results = [results_tfidf, results_bm25, results_random]
question_collection = questions_for_ranking
passages_collection = passages_df
folder_path = 'data/results_ranking'

save_results(results, question_collection, passages_collection, folder_path)
