==> notebooks/2_ranking_text_generation_pipeline/2_2_ranking_all_questions_nonfact_filtering.ipynb <==
### Setup CD

%cd  /Users/natalipeeva/Documents/GitHub/Automatic-Answering-of-City-Council-Questions/

import sys
import os
src_dir = os.path.join(os.getcwd(), 'src_clean')
sys.path.append(src_dir)

import pandas as pd
from hashids import Hashids
from tabulate import tabulate

from src_clean.preprocessing.passages import create_passages_dataframe, add_passages_ids
from src_clean.preprocessing.text_preprocessing import preprocess_text
from src_clean.ranking.sparse_ranking import perform_tfidf_search, perform_random_search, perform_bm25_search
from src_clean.ranking.evaluation import calculate_average_metrics_retrieval, calculate_metrics_answer_similarity, simulate_answer

### Read Data

questions = pd.read_csv('data/question_answer/questions.csv')


len(questions)

questions = pd.read_csv('data/question_answer/questions.csv')
collection = pd.read_csv('data/amsterdam/amsterdam_full.csv')

## Prepare Data for ranking

### Create passages

passages_df = create_passages_dataframe(collection) # 50 secs for execution
len(passages_df)

### Clean passages

#### Drop duplicates in general; drop duplictates in column 'Textual_Content'

passages_df = passages_df.drop_duplicates() # drop dup.
len(passages_df)

passages_df = passages_df.drop_duplicates(subset='Textual_Content', keep=False) # drop dup. passages
len(passages_df)

#### Pre-process Passages Text

passages_df['Preprocessed_Text'] = passages_df['Textual_Content'].apply(lambda x: preprocess_text(x,stem=True,
                                                                                      remove_stopwords=True,
                                                                                      lowercase_text=True,
                                                                                      remove_punct=True)) # 40 seconds

#### Add Passage ids

hashids = Hashids()
passages_df["id"] = [hashids.encode(i) for i in range(len(passages_df))]

passages_df.tail()

### Clean Questions

#### Pre-process questions and answers

questions.dropna(subset=['Question', 'Answer'], inplace=True)

questions = questions.drop_duplicates()

questions.drop_duplicates(subset='Answer', inplace=True)

questions.drop_duplicates(subset='Question', inplace=True)

questions.head()

questions['Preprocessed_Question'] = questions['Question'].apply(lambda x: preprocess_text(x,stem=True,
                                                                                      remove_stopwords=True,
                                                                                      lowercase_text=True,
                                                                                      remove_punct=True)) 

questions['Preprocessed_Answer'] = questions['Answer'].apply(lambda x: preprocess_text(x,stem=True,
                                                                                      remove_stopwords=True,
                                                                                      lowercase_text=True,
                                                                                      remove_punct=True)) 

questions.drop_duplicates(subset='Preprocessed_Question', inplace=True)

questions.drop_duplicates(subset='Preprocessed_Answer', inplace=True)

questions.describe()

questions=questions[questions['Preprocessed_Answer'].str.split().str.len() > 10]

words_to_filter = ['bekend', 'kennisgenomen', 'op de hoogte', 'standpunt', 'mening', 'onbekend', 'vindt', 'visie',
                   'herkent']

questions = questions[~questions['Answer'].str.split().apply(lambda x: any(word.lower() in x for word in words_to_filter))]

questions = questions[~questions['Question'].str.split().apply(lambda x: any(word.lower() in x for word in words_to_filter))]

len(questions)

questions=questions[questions['Preprocessed_Answer'].str.split().str.len() < 400]

#questions.head()

##### Add question ids 

hashids = Hashids()
questions["question_id"] = [hashids.encode(i) for i in range(len(questions))]

# questions.tail() # check

#### prepare dataset

from datasets import Dataset
import pyarrow as pa

arrow_table = pa.Table.from_pandas(questions)
arrow_dict = arrow_table.to_pydict()
questions = Dataset.from_dict(arrow_dict)

# questions[0] # check

arrow_table = pa.Table.from_pandas(passages_df)
arrow_dict = arrow_table.to_pydict()
passages_df = Dataset.from_dict(arrow_dict)

# passages_df[1000] # check

type(passages_df)

## Ranking

### TF-IDF search

results_tfidf = perform_tfidf_search(questions, passages_df, k=100) # 6 mins for whole collection

### Random

import numpy as np

def random_search(query, collection, k, get_true_passages=False):
    if get_true_passages:
        true_documents = query['passages_ids']
    else:
        true_documents = []

    # Randomly select k documents from the collection
    collection_indices = np.arange(len(collection))
    random_indices = np.random.choice(collection_indices, size=k, replace=False)
    random_results = [collection[int(i)] for i in random_indices]

    ranked_ids = [result['id'] for result in random_results]
    ranked_text = [result['Textual_Content'] for result in random_results]
    ranked_preprocessed = [result['Preprocessed_Text'] for result in random_results]
    scores = []  # Assign random scores for demonstration purposes

    # Dictionary with search results
    search_results = {
        'question_id': query['question_id'],
        'question': query['Question'],
        'ranked_ids': ranked_ids,
        'ranked_text_preprocessed': ranked_preprocessed,
        'ranked_text': ranked_text,
        'true_passages': true_documents,
        'scores': scores,
        'answer': query['Answer'],
        'preprocessed_question': query['Preprocessed_Question'],
        'preprocessed_answer': query['Preprocessed_Answer']
    }

    return search_results


def perform_random_search(queries, collection, k, get_true_passages= False):
    """
    Perform random retrieval search for each query in a list of queries.
    Input:
        queries - a list of queries
        collection: a pandas DataFrame representing the collection
        k: the number of top search results to retrieve
    Output: a list of dictionaries containing the search results
    """
    search_results = []

    for query in queries:
        results = random_search(query, collection, k=k, get_true_passages=get_true_passages)
        search_results.append(results)

    return search_results

results_random = perform_random_search(questions, passages_df, k=100) # 18 secs

## Save ranked results

import os
import pickle

def save_results(results, question_collection, passages_collection, folder_path):
    # Create the folder if it doesn't exist
    os.makedirs(folder_path, exist_ok=True)
    
    # Get the variable names of question_collection and passages_collection
    question_var_name = [var_name for var_name, var_val in globals().items() if var_val is question_collection][0]
    passages_var_name = [var_name for var_name, var_val in globals().items() if var_val is passages_collection][0]

    # Save each result variable with its corresponding filename
    for result in results:
        # Get the variable name of the result variable
        result_var_name = [var_name for var_name, var_val in globals().items() if var_val is result][0]
        
        # Generate the filename based on variable names
        filename = '{}_{}_{}.pickle'.format(result_var_name, question_var_name, passages_var_name)
        
        file_path = os.path.join(folder_path, filename)
        with open(file_path, 'wb') as f:
            pickle.dump(result, f)


results = [results_tfidf, results_random]
question_collection = questions
passages_collection = passages_df
folder_path = 'data/results_ranking'

save_results(results, question_collection, passages_collection, folder_path)

import pickle
with open('data/results_ranking/questions_preprocessed.pickle', 'wb') as f:
    pickle.dump(questions, f)

with open('data/results_ranking/passages_preprocessed.pickle', 'wb') as f:
    pickle.dump(passages_df, f)

import pickle

with open('/Users/natalipeeva/Documents/GitHub/Automatic-Answering-of-City-Council-Questions/data/results_ranking/results_tfidf_questions_passages_df.pickle', 'rb') as f:
    test = pickle.load(f)
