==> notebooks/2_ranking_text_generation_pipeline/3_1_get_train_test_set.ipynb <==
import pickle
import pandas as pd

import os

%cd '/Users/natalipeeva/Documents/GitHub/Automatic-Answering-of-City-Council-Questions/'

#### Remove amsetrdam questions from training set

with open ('data/results_ranking/with-nonfactual/questions_preprocessed.pickle', 'rb') as f:
    preprocessed_questions = pickle.load(f)

questions_for_ranking = pd.read_csv('data/question_answer/amsterdam_questions.csv')

good_samples = []
for sample in preprocessed_questions:
    if sample['Question'] in list(questions_for_ranking['Question']):
        good_samples.append(sample)

train_dataset = []
for sample in preprocessed_questions:
    if not sample['Question'] in list(questions_for_ranking['Question']):
        train_dataset.append(sample)

from datasets import Dataset
df = pd.DataFrame(train_dataset)
train_dataset = Dataset.from_pandas(df)

train_dataset

#### Load passages preprocessed

with open('data/results_ranking/passages_preprocessed.pickle', 'rb') as f:
    passages_preprocessed = pickle.load(f)

type(passages_preprocessed)

#### Ranking

from src_clean.ranking.sparse_ranking import perform_tfidf_search, perform_random_search
from src_clean.ranking.evaluation import calculate_metrics_answer_similarity, simulate_answer

#### TFIDF

results_tfidf = perform_tfidf_search(train_dataset, passages_preprocessed, k=10) # 6 mins for whole collection

with open ('data/results_ranking/with-nonfactual/training_sample_tfidf.pickle', 'wb') as f:
    pickle.dump(results_tfidf,f)
